switch to uint64_t for portability

+++ ai.c
@@ -182,7 +182,7 @@ ai_brain_save(ai_brain_t *brain, const char *filename)
 
   f = fopen(filename, "wb");
 
-  fwrite(&brain->layer_count, sizeof(size_t), 1, f);
+  fwrite(&brain->layer_count, sizeof(uint64_t), 1, f);
   for (i = 0; i < brain->layer_count; i++) {
     l = brain->layers[i];
     fwrite(&l.activation.type, sizeof(ai_activation_type_t), 1, f);
@@ -194,8 +194,8 @@ ai_brain_save(ai_brain_t *brain, const char *filename)
       fwrite(&l.activation.data.softmax, sizeof(ai_activation_data_softmax), 1, f);
       break;
     }
-    fwrite(&l.input_size, sizeof(size_t), 1, f);
-    fwrite(&l.output_size, sizeof(size_t), 1, f);
+    fwrite(&l.input_size, sizeof(uint64_t), 1, f);
+    fwrite(&l.output_size, sizeof(uint64_t), 1, f);
     fwrite(l.weights, sizeof(float), l.input_size * l.output_size, f);
     fwrite(l.biases, sizeof(float), l.output_size, f);
   }
@@ -214,7 +214,7 @@ ai_brain_load(ai_brain_t *brain, const char *filename)
 
   f = fopen(filename, "rb");
 
-  fread(&brain->layer_count, sizeof(size_t), 1, f);
+  fread(&brain->layer_count, sizeof(uint64_t), 1, f);
   brain->layers = malloc(sizeof(ai_layer_t) * brain->layer_count);
   for (i = 0; i < brain->layer_count; i++) {
     l = &brain->layers[i];
@@ -227,8 +227,8 @@ ai_brain_load(ai_brain_t *brain, const char *filename)
       fread(&l->activation.data.softmax, sizeof(ai_activation_data_softmax), 1, f);
       break;
     }
-    fread(&l->input_size, sizeof(size_t), 1, f);
-    fread(&l->output_size, sizeof(size_t), 1, f);
+    fread(&l->input_size, sizeof(uint64_t), 1, f);
+    fread(&l->output_size, sizeof(uint64_t), 1, f);
     l->weights = malloc(sizeof(float) * l->input_size * l->output_size);
     l->biases = malloc(sizeof(float) * l->output_size);
     l->outputs = malloc(sizeof(float) * l->output_size);
